1) Cosmetics in client-side transactions 2) Server meta for tables and files 3) NFileClient::TFileWriter 4) Cypress experiments used TTransactionManager and TFileWriter

+++ yt/ytlib/table_client/chunk_writer.h
@@ -27,7 +27,7 @@ public:
 
     struct TConfig
     {
-        int BlockSize;
+        i64 BlockSize;
 
         TConfig()
             // ToDo: make configurable
